pyproject.toml
[project]
name = "rag-plus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "datasets>=3.6.0",
    "dspy-ai>=2.6.27",
    "ipykernel>=6.29.5",
    "numpy>=2.3.0",
    "openai>=1.87.0",
    "pip>=25.1.1",
    "tqdm>=4.67.1",
]
